+++ tests/metrics/functional/test_classification.py
@@ -25,18 +25,15 @@ from pytorch_lightning.metrics.functional.classification import (
 from pytorch_lightning.metrics.utils import to_onehot, get_num_classes, to_categorical
 
 
-@pytest.mark.parametrize(
-    ["sklearn_metric", "torch_metric", "only_binary"],
-    [
-        pytest.param(partial(sk_jaccard_score, average="macro"), iou, False, id="iou"),
-        pytest.param(partial(sk_precision, average="micro"), precision, False, id="precision"),
-        pytest.param(partial(sk_recall, average="micro"), recall, False, id="recall"),
-        pytest.param(sk_roc_auc_score, auroc, True, id="auroc"),
-    ],
-)
+@pytest.mark.parametrize(['sklearn_metric', 'torch_metric', 'only_binary'], [
+    pytest.param(partial(sk_jaccard_score, average='macro'), iou, False, id='iou'),
+    pytest.param(partial(sk_precision, average='micro'), precision, False, id='precision'),
+    pytest.param(partial(sk_recall, average='micro'), recall, False, id='recall'),
+    pytest.param(sk_roc_auc_score, auroc, True, id='auroc')
+])
 def test_against_sklearn(sklearn_metric, torch_metric, only_binary):
     """Compare PL metrics to sklearn version. """
-    device = "cuda" if torch.cuda.is_available() else "cpu"
+    device = 'cuda' if torch.cuda.is_available() else 'cpu'
 
     # for metrics with only_binary=False, we try out different combinations of number
     # of labels in pred and target (also test binary)
@@ -47,7 +44,8 @@ def test_against_sklearn(sklearn_metric, torch_metric, only_binary):
         pred = torch.randint(n_cls_pred, (300,), device=device)
         target = torch.randint(n_cls_target, (300,), device=device)
 
-        sk_score = sklearn_metric(target.cpu().detach().numpy(), pred.cpu().detach().numpy())
+        sk_score = sklearn_metric(target.cpu().detach().numpy(),
+                                  pred.cpu().detach().numpy())
         pl_score = torch_metric(pred, target)
 
         # if multi output
@@ -60,21 +58,20 @@ def test_against_sklearn(sklearn_metric, torch_metric, only_binary):
             assert torch.allclose(sk_score, pl_score)
 
 
-@pytest.mark.parametrize("class_reduction", ["micro", "macro", "weighted"])
-@pytest.mark.parametrize(
-    ["sklearn_metric", "torch_metric"],
-    [
-        pytest.param(sk_precision, precision, id="precision"),
-        pytest.param(sk_recall, recall, id="recall"),
-    ],
-)
+@pytest.mark.parametrize('class_reduction', ['micro', 'macro', 'weighted'])
+@pytest.mark.parametrize(['sklearn_metric', 'torch_metric'], [
+    pytest.param(sk_precision, precision, id='precision'),
+    pytest.param(sk_recall, recall, id='recall'),
+])
 def test_different_reduction_against_sklearn(class_reduction, sklearn_metric, torch_metric):
-    """Test metrics where the class_reduction parameter have a correponding
-    value in sklearn"""
-    device = "cuda" if torch.cuda.is_available() else "cpu"
+    """ Test metrics where the class_reduction parameter have a correponding
+        value in sklearn """
+    device = 'cuda' if torch.cuda.is_available() else 'cpu'
     pred = torch.randint(10, (300,), device=device)
     target = torch.randint(10, (300,), device=device)
-    sk_score = sklearn_metric(target.cpu().detach().numpy(), pred.cpu().detach().numpy(), average=class_reduction)
+    sk_score = sklearn_metric(target.cpu().detach().numpy(),
+                              pred.cpu().detach().numpy(),
+                              average=class_reduction)
     sk_score = torch.tensor(sk_score, dtype=torch.float, device=device)
     pl_score = torch_metric(pred, target, class_reduction=class_reduction)
     assert torch.allclose(sk_score, pl_score)
@@ -82,12 +79,10 @@ def test_different_reduction_against_sklearn(class_reduction, sklearn_metric, to
 
 def test_onehot():
     test_tensor = torch.tensor([[0, 1, 2, 3, 4], [5, 6, 7, 8, 9]])
-    expected = torch.stack(
-        [
-            torch.cat([torch.eye(5, dtype=int), torch.zeros((5, 5), dtype=int)]),
-            torch.cat([torch.zeros((5, 5), dtype=int), torch.eye(5, dtype=int)]),
-        ]
-    )
+    expected = torch.stack([
+        torch.cat([torch.eye(5, dtype=int), torch.zeros((5, 5), dtype=int)]),
+        torch.cat([torch.zeros((5, 5), dtype=int), torch.eye(5, dtype=int)])
+    ])
 
     assert test_tensor.shape == (2, 5)
     assert expected.shape == (2, 10, 5)
@@ -105,12 +100,10 @@ def test_onehot():
 
 
 def test_to_categorical():
-    test_tensor = torch.stack(
-        [
-            torch.cat([torch.eye(5, dtype=int), torch.zeros((5, 5), dtype=int)]),
-            torch.cat([torch.zeros((5, 5), dtype=int), torch.eye(5, dtype=int)]),
-        ]
-    ).to(torch.float)
+    test_tensor = torch.stack([
+        torch.cat([torch.eye(5, dtype=int), torch.zeros((5, 5), dtype=int)]),
+        torch.cat([torch.zeros((5, 5), dtype=int), torch.eye(5, dtype=int)])
+    ]).to(torch.float)
 
     expected = torch.tensor([[0, 1, 2, 3, 4], [5, 6, 7, 8, 9]])
     assert expected.shape == (2, 5)
@@ -122,25 +115,20 @@ def test_to_categorical():
     assert torch.allclose(result, expected.to(result.dtype))
 
 
-@pytest.mark.parametrize(
-    ["pred", "target", "num_classes", "expected_num_classes"],
-    [
-        pytest.param(torch.rand(32, 10, 28, 28), torch.randint(10, (32, 28, 28)), 10, 10),
-        pytest.param(torch.rand(32, 10, 28, 28), torch.randint(10, (32, 28, 28)), None, 10),
-        pytest.param(torch.rand(32, 28, 28), torch.randint(10, (32, 28, 28)), None, 10),
-    ],
-)
+@pytest.mark.parametrize(['pred', 'target', 'num_classes', 'expected_num_classes'], [
+    pytest.param(torch.rand(32, 10, 28, 28), torch.randint(10, (32, 28, 28)), 10, 10),
+    pytest.param(torch.rand(32, 10, 28, 28), torch.randint(10, (32, 28, 28)), None, 10),
+    pytest.param(torch.rand(32, 28, 28), torch.randint(10, (32, 28, 28)), None, 10),
+])
 def test_get_num_classes(pred, target, num_classes, expected_num_classes):
     assert get_num_classes(pred, target, num_classes) == expected_num_classes
 
 
-@pytest.mark.parametrize(
-    ["pred", "target", "expected_tp", "expected_fp", "expected_tn", "expected_fn", "expected_support"],
-    [
-        pytest.param(torch.tensor([0.0, 2.0, 4.0, 4.0]), torch.tensor([0.0, 4.0, 3.0, 4.0]), 1, 1, 1, 1, 2),
-        pytest.param(to_onehot(torch.tensor([0.0, 2.0, 4.0, 4.0])), torch.tensor([0.0, 4.0, 3.0, 4.0]), 1, 1, 1, 1, 2),
-    ],
-)
+@pytest.mark.parametrize(['pred', 'target', 'expected_tp', 'expected_fp',
+                          'expected_tn', 'expected_fn', 'expected_support'], [
+    pytest.param(torch.tensor([0., 2., 4., 4.]), torch.tensor([0., 4., 3., 4.]), 1, 1, 1, 1, 2),
+    pytest.param(to_onehot(torch.tensor([0., 2., 4., 4.])), torch.tensor([0., 4., 3., 4.]), 1, 1, 1, 1, 2)
+])
 def test_stat_scores(pred, target, expected_tp, expected_fp, expected_tn, expected_fn, expected_support):
     tp, fp, tn, fn, sup = stat_scores(pred, target, class_index=4)
 
@@ -151,54 +139,18 @@ def test_stat_scores(pred, target, expected_tp, expected_fp, expected_tn, expect
     assert sup.item() == expected_support
 
 
-@pytest.mark.parametrize(
-    ["pred", "target", "reduction", "expected_tp", "expected_fp", "expected_tn", "expected_fn", "expected_support"],
-    [
-        pytest.param(
-            torch.tensor([0.0, 2.0, 4.0, 4.0]),
-            torch.tensor([0.0, 4.0, 3.0, 4.0]),
-            "none",
-            [1, 0, 0, 0, 1],
-            [0, 0, 1, 0, 1],
-            [3, 4, 3, 3, 1],
-            [0, 0, 0, 1, 1],
-            [1, 0, 0, 1, 2],
-        ),
-        pytest.param(
-            to_onehot(torch.tensor([0.0, 2.0, 4.0, 4.0])),
-            torch.tensor([0.0, 4.0, 3.0, 4.0]),
-            "none",
-            [1, 0, 0, 0, 1],
-            [0, 0, 1, 0, 1],
-            [3, 4, 3, 3, 1],
-            [0, 0, 0, 1, 1],
-            [1, 0, 0, 1, 2],
-        ),
-        pytest.param(
-            to_onehot(torch.tensor([0.0, 2.0, 4.0, 4.0])),
-            torch.tensor([0.0, 4.0, 3.0, 4.0]),
-            "sum",
-            torch.tensor(2),
-            torch.tensor(2),
-            torch.tensor(14),
-            torch.tensor(2),
-            torch.tensor(4),
-        ),
-        pytest.param(
-            to_onehot(torch.tensor([0.0, 2.0, 4.0, 4.0])),
-            torch.tensor([0.0, 4.0, 3.0, 4.0]),
-            "elementwise_mean",
-            torch.tensor(0.4),
-            torch.tensor(0.4),
-            torch.tensor(2.8),
-            torch.tensor(0.4),
-            torch.tensor(0.8),
-        ),
-    ],
-)
-def test_stat_scores_multiclass(
-    pred, target, reduction, expected_tp, expected_fp, expected_tn, expected_fn, expected_support
-):
+@pytest.mark.parametrize(['pred', 'target', 'reduction', 'expected_tp', 'expected_fp',
+                          'expected_tn', 'expected_fn', 'expected_support'], [
+    pytest.param(torch.tensor([0., 2., 4., 4.]), torch.tensor([0., 4., 3., 4.]), 'none',
+                 [1, 0, 0, 0, 1], [0, 0, 1, 0, 1], [3, 4, 3, 3, 1], [0, 0, 0, 1, 1], [1, 0, 0, 1, 2]),
+    pytest.param(to_onehot(torch.tensor([0., 2., 4., 4.])), torch.tensor([0., 4., 3., 4.]), 'none',
+                 [1, 0, 0, 0, 1], [0, 0, 1, 0, 1], [3, 4, 3, 3, 1], [0, 0, 0, 1, 1], [1, 0, 0, 1, 2]),
+    pytest.param(to_onehot(torch.tensor([0., 2., 4., 4.])), torch.tensor([0., 4., 3., 4.]), 'sum',
+                 torch.tensor(2), torch.tensor(2), torch.tensor(14), torch.tensor(2), torch.tensor(4)),
+    pytest.param(to_onehot(torch.tensor([0., 2., 4., 4.])), torch.tensor([0., 4., 3., 4.]), 'elementwise_mean',
+                 torch.tensor(0.4), torch.tensor(0.4), torch.tensor(2.8), torch.tensor(0.4), torch.tensor(0.8))
+])
+def test_stat_scores_multiclass(pred, target, reduction, expected_tp, expected_fp, expected_tn, expected_fn, expected_support):
     tp, fp, tn, fn, sup = stat_scores_multiple_classes(pred, target, reduction=reduction)
 
     assert torch.allclose(torch.tensor(expected_tp).to(tp), tp)
@@ -208,30 +160,22 @@ def test_stat_scores_multiclass(
     assert torch.allclose(torch.tensor(expected_support).to(sup), sup)
 
 
-@pytest.mark.parametrize(
-    ["pred", "target", "expected_prec", "expected_rec"],
-    [
-        pytest.param(torch.tensor([1.0, 0.0, 1.0, 0.0]), torch.tensor([0.0, 1.0, 1.0, 0.0]), [0.5, 0.5], [0.5, 0.5]),
-        pytest.param(
-            to_onehot(torch.tensor([1.0, 0.0, 1.0, 0.0])), torch.tensor([0.0, 1.0, 1.0, 0.0]), [0.5, 0.5], [0.5, 0.5]
-        ),
-    ],
-)
+@pytest.mark.parametrize(['pred', 'target', 'expected_prec', 'expected_rec'], [
+    pytest.param(torch.tensor([1., 0., 1., 0.]), torch.tensor([0., 1., 1., 0.]), [0.5, 0.5], [0.5, 0.5]),
+    pytest.param(to_onehot(torch.tensor([1., 0., 1., 0.])), torch.tensor([0., 1., 1., 0.]), [0.5, 0.5], [0.5, 0.5])
+])
 def test_precision_recall(pred, target, expected_prec, expected_rec):
-    prec = precision(pred, target, class_reduction="none")
-    rec = recall(pred, target, class_reduction="none")
+    prec = precision(pred, target, class_reduction='none')
+    rec = recall(pred, target, class_reduction='none')
 
     assert torch.allclose(torch.tensor(expected_prec).to(prec), prec)
     assert torch.allclose(torch.tensor(expected_rec).to(rec), rec)
 
 
-@pytest.mark.parametrize(
-    ["sample_weight", "pos_label", "exp_shape"],
-    [
-        pytest.param(1, 1.0, 42),
-        pytest.param(None, 1.0, 42),
-    ],
-)
+@pytest.mark.parametrize(['sample_weight', 'pos_label', "exp_shape"], [
+    pytest.param(1, 1., 42),
+    pytest.param(None, 1., 42),
+])
 def test_binary_clf_curve(sample_weight, pos_label, exp_shape):
     # TODO: move back the pred and target to test func arguments
     #  if you fix the array inside the function, you'd also have fix the shape,
@@ -252,94 +196,90 @@ def test_binary_clf_curve(sample_weight, pos_label, exp_shape):
     assert thresh.shape == (exp_shape,)
 
 
-@pytest.mark.parametrize(
-    ["pred", "target", "expected"],
-    [
-        pytest.param([0, 1, 0, 1], [0, 1, 0, 1], 1.0),
-        pytest.param([1, 1, 0, 0], [0, 0, 1, 1], 0.0),
-        pytest.param([1, 1, 1, 1], [1, 1, 0, 0], 0.5),
-        pytest.param([1, 1, 0, 0], [1, 1, 0, 0], 1.0),
-        pytest.param([0.5, 0.5, 0.5, 0.5], [1, 1, 0, 0], 0.5),
-    ],
-)
+@pytest.mark.parametrize(['pred', 'target', 'expected'], [
+    pytest.param([0, 1, 0, 1], [0, 1, 0, 1], 1.),
+    pytest.param([1, 1, 0, 0], [0, 0, 1, 1], 0.),
+    pytest.param([1, 1, 1, 1], [1, 1, 0, 0], 0.5),
+    pytest.param([1, 1, 0, 0], [1, 1, 0, 0], 1.),
+    pytest.param([0.5, 0.5, 0.5, 0.5], [1, 1, 0, 0], 0.5),
+])
 def test_auroc(pred, target, expected):
     score = auroc(torch.tensor(pred), torch.tensor(target)).item()
     assert score == expected
 
 
 def test_multiclass_auroc():
-    with pytest.raises(ValueError, match=r".*probabilities, i.e. they should sum up to 1.0 over classes"):
-        _ = multiclass_auroc(pred=torch.tensor([[0.9, 0.9], [1.0, 0]]), target=torch.tensor([0, 1]))
+    with pytest.raises(ValueError,
+                       match=r".*probabilities, i.e. they should sum up to 1.0 over classes"):
+        _ = multiclass_auroc(pred=torch.tensor([[0.9, 0.9],
+                                                [1.0, 0]]),
+                             target=torch.tensor([0, 1]))
 
-    with pytest.raises(ValueError, match=r".*not defined when all of the classes do not occur in the target.*"):
-        _ = multiclass_auroc(pred=torch.rand((4, 3)).softmax(dim=1), target=torch.tensor([1, 0, 1, 0]))
+    with pytest.raises(ValueError,
+                       match=r".*not defined when all of the classes do not occur in the target.*"):
+        _ = multiclass_auroc(pred=torch.rand((4, 3)).softmax(dim=1),
+                             target=torch.tensor([1, 0, 1, 0]))
 
-    with pytest.raises(ValueError, match=r".*does not equal the number of classes passed in 'num_classes'.*"):
-        _ = multiclass_auroc(
-            pred=torch.rand((5, 4)).softmax(dim=1), target=torch.tensor([0, 1, 2, 2, 3]), num_classes=6
-        )
+    with pytest.raises(ValueError,
+                       match=r".*does not equal the number of classes passed in 'num_classes'.*"):
+        _ = multiclass_auroc(pred=torch.rand((5, 4)).softmax(dim=1),
+                             target=torch.tensor([0, 1, 2, 2, 3]),
+                             num_classes=6)
 
 
-@pytest.mark.parametrize("n_cls", [2, 5, 10, 50])
+@pytest.mark.parametrize('n_cls', [2, 5, 10, 50])
 def test_multiclass_auroc_against_sklearn(n_cls):
-    device = "cuda" if torch.cuda.is_available() else "cpu"
+    device = 'cuda' if torch.cuda.is_available() else 'cpu'
 
     n_samples = 300
     pred = torch.rand(n_samples, n_cls, device=device).softmax(dim=1)
     target = torch.randint(n_cls, (n_samples,), device=device)
     # Make sure target includes all class labels so that multiclass AUROC is defined
-    target[10 : 10 + n_cls] = torch.arange(n_cls)
+    target[10:10 + n_cls] = torch.arange(n_cls)
 
     pl_score = multiclass_auroc(pred, target)
     # For the binary case, sklearn expects an (n_samples,) array of probabilities of
     # the positive class
     pred = pred[:, 1] if n_cls == 2 else pred
-    sk_score = sk_roc_auc_score(target.cpu().detach().numpy(), pred.cpu().detach().numpy(), multi_class="ovr")
+    sk_score = sk_roc_auc_score(target.cpu().detach().numpy(),
+                                pred.cpu().detach().numpy(),
+                                multi_class="ovr")
 
     sk_score = torch.tensor(sk_score, dtype=torch.float, device=device)
     assert torch.allclose(sk_score, pl_score)
 
 
-@pytest.mark.parametrize(
-    ["x", "y", "expected"],
-    [
-        pytest.param([0, 1], [0, 1], 0.5),
-        pytest.param([1, 0], [0, 1], 0.5),
-        pytest.param([1, 0, 0], [0, 1, 1], 0.5),
-        pytest.param([0, 1], [1, 1], 1),
-        pytest.param([0, 0.5, 1], [0, 0.5, 1], 0.5),
-    ],
-)
+@pytest.mark.parametrize(['x', 'y', 'expected'], [
+    pytest.param([0, 1], [0, 1], 0.5),
+    pytest.param([1, 0], [0, 1], 0.5),
+    pytest.param([1, 0, 0], [0, 1, 1], 0.5),
+    pytest.param([0, 1], [1, 1], 1),
+    pytest.param([0, 0.5, 1], [0, 0.5, 1], 0.5),
+])
 def test_auc(x, y, expected):
     # Test Area Under Curve (AUC) computation
     assert auc(torch.tensor(x), torch.tensor(y)) == expected
 
 
-@pytest.mark.parametrize(
-    ["pred", "target", "expected"],
-    [
-        pytest.param([[0, 0], [1, 1]], [[0, 0], [1, 1]], 1.0),
-        pytest.param([[1, 1], [0, 0]], [[0, 0], [1, 1]], 0.0),
-        pytest.param([[1, 1], [1, 1]], [[1, 1], [0, 0]], 2 / 3),
-        pytest.param([[1, 1], [0, 0]], [[1, 1], [0, 0]], 1.0),
-    ],
-)
+@pytest.mark.parametrize(['pred', 'target', 'expected'], [
+    pytest.param([[0, 0], [1, 1]], [[0, 0], [1, 1]], 1.),
+    pytest.param([[1, 1], [0, 0]], [[0, 0], [1, 1]], 0.),
+    pytest.param([[1, 1], [1, 1]], [[1, 1], [0, 0]], 2 / 3),
+    pytest.param([[1, 1], [0, 0]], [[1, 1], [0, 0]], 1.),
+])
 def test_dice_score(pred, target, expected):
     score = dice_score(torch.tensor(pred), torch.tensor(target))
     assert score == expected
 
 
-@pytest.mark.parametrize(
-    ["half_ones", "reduction", "ignore_index", "expected"],
-    [
-        pytest.param(False, "none", None, torch.Tensor([1, 1, 1])),
-        pytest.param(False, "elementwise_mean", None, torch.Tensor([1])),
-        pytest.param(False, "none", 0, torch.Tensor([1, 1])),
-        pytest.param(True, "none", None, torch.Tensor([0.5, 0.5, 0.5])),
-        pytest.param(True, "elementwise_mean", None, torch.Tensor([0.5])),
-        pytest.param(True, "none", 0, torch.Tensor([0.5, 0.5])),
-    ],
-)
+@pytest.mark.parametrize(['half_ones', 'reduction', 'ignore_index', 'expected'], [
+    pytest.param(False, 'none', None, torch.Tensor([1, 1, 1])),
+    pytest.param(False, 'elementwise_mean', None, torch.Tensor([1])),
+    pytest.param(False, 'none', 0, torch.Tensor([1, 1])),
+    pytest.param(True, 'none', None, torch.Tensor([0.5, 0.5, 0.5])),
+    pytest.param(True, 'elementwise_mean', None, torch.Tensor([0.5])),
+    pytest.param(True, 'none', 0, torch.Tensor([0.5, 0.5])),
+])
 def test_iou(half_ones, reduction, ignore_index, expected):
     pred = (torch.arange(120) % 3).view(-1, 1)
     target = (torch.arange(120) % 3).view(-1, 1)
@@ -356,17 +296,19 @@ def test_iou(half_ones, reduction, ignore_index, expected):
 
 def test_iou_input_check():
     with pytest.raises(ValueError, match=r"'pred' shape (.*) must equal 'target' shape (.*)"):
-        _ = iou(pred=torch.randint(0, 2, (3, 4, 3)), target=torch.randint(0, 2, (3, 3)))
+        _ = iou(pred=torch.randint(0, 2, (3, 4, 3)),
+                target=torch.randint(0, 2, (3, 3)))
 
     with pytest.raises(ValueError, match="'pred' must contain integer targets."):
-        _ = iou(pred=torch.rand((3, 3)), target=torch.randint(0, 2, (3, 3)))
+        _ = iou(pred=torch.rand((3, 3)),
+                target=torch.randint(0, 2, (3, 3)))
 
 
-@pytest.mark.parametrize("metric", [auroc])
+@pytest.mark.parametrize('metric', [auroc])
 def test_error_on_multiclass_input(metric):
     """ check that these metrics raise an error if they are used for multiclass problems  """
-    pred = torch.randint(0, 10, (100,))
-    target = torch.randint(0, 10, (100,))
+    pred = torch.randint(0, 10, (100, ))
+    target = torch.randint(0, 10, (100, ))
     with pytest.raises(ValueError, match="AUROC metric is meant for binary classification"):
         _ = metric(pred, target)
 
@@ -374,38 +316,35 @@ def test_error_on_multiclass_input(metric):
 # TODO: When the jaccard_score of the sklearn version we use accepts `zero_division` (see
 #       https://github.com/scikit-learn/scikit-learn/pull/17866), consider adding a test here against our
 #       `absent_score`.
-@pytest.mark.parametrize(
-    ["pred", "target", "ignore_index", "absent_score", "num_classes", "expected"],
-    [
-        # Note that -1 is used as the absent_score in almost all tests here to distinguish it from the range of valid
-        # scores the function can return ([0., 1.] range, inclusive).
-        # 2 classes, class 0 is correct everywhere, class 1 is absent.
-        pytest.param([0], [0], None, -1.0, 2, [1.0, -1.0]),
-        pytest.param([0, 0], [0, 0], None, -1.0, 2, [1.0, -1.0]),
-        # absent_score not applied if only class 0 is present and it's the only class.
-        pytest.param([0], [0], None, -1.0, 1, [1.0]),
-        # 2 classes, class 1 is correct everywhere, class 0 is absent.
-        pytest.param([1], [1], None, -1.0, 2, [-1.0, 1.0]),
-        pytest.param([1, 1], [1, 1], None, -1.0, 2, [-1.0, 1.0]),
-        # When 0 index ignored, class 0 does not get a score (not even the absent_score).
-        pytest.param([1], [1], 0, -1.0, 2, [1.0]),
-        # 3 classes. Only 0 and 2 are present, and are perfectly predicted. 1 should get absent_score.
-        pytest.param([0, 2], [0, 2], None, -1.0, 3, [1.0, -1.0, 1.0]),
-        pytest.param([2, 0], [2, 0], None, -1.0, 3, [1.0, -1.0, 1.0]),
-        # 3 classes. Only 0 and 1 are present, and are perfectly predicted. 2 should get absent_score.
-        pytest.param([0, 1], [0, 1], None, -1.0, 3, [1.0, 1.0, -1.0]),
-        pytest.param([1, 0], [1, 0], None, -1.0, 3, [1.0, 1.0, -1.0]),
-        # 3 classes, class 0 is 0.5 IoU, class 1 is 0 IoU (in pred but not target; should not get absent_score), class
-        # 2 is absent.
-        pytest.param([0, 1], [0, 0], None, -1.0, 3, [0.5, 0.0, -1.0]),
-        # 3 classes, class 0 is 0.5 IoU, class 1 is 0 IoU (in target but not pred; should not get absent_score), class
-        # 2 is absent.
-        pytest.param([0, 0], [0, 1], None, -1.0, 3, [0.5, 0.0, -1.0]),
-        # Sanity checks with absent_score of 1.0.
-        pytest.param([0, 2], [0, 2], None, 1.0, 3, [1.0, 1.0, 1.0]),
-        pytest.param([0, 2], [0, 2], 0, 1.0, 3, [1.0, 1.0]),
-    ],
-)
+@pytest.mark.parametrize(['pred', 'target', 'ignore_index', 'absent_score', 'num_classes', 'expected'], [
+    # Note that -1 is used as the absent_score in almost all tests here to distinguish it from the range of valid
+    # scores the function can return ([0., 1.] range, inclusive).
+    # 2 classes, class 0 is correct everywhere, class 1 is absent.
+    pytest.param([0], [0], None, -1., 2, [1., -1.]),
+    pytest.param([0, 0], [0, 0], None, -1., 2, [1., -1.]),
+    # absent_score not applied if only class 0 is present and it's the only class.
+    pytest.param([0], [0], None, -1., 1, [1.]),
+    # 2 classes, class 1 is correct everywhere, class 0 is absent.
+    pytest.param([1], [1], None, -1., 2, [-1., 1.]),
+    pytest.param([1, 1], [1, 1], None, -1., 2, [-1., 1.]),
+    # When 0 index ignored, class 0 does not get a score (not even the absent_score).
+    pytest.param([1], [1], 0, -1., 2, [1.0]),
+    # 3 classes. Only 0 and 2 are present, and are perfectly predicted. 1 should get absent_score.
+    pytest.param([0, 2], [0, 2], None, -1., 3, [1., -1., 1.]),
+    pytest.param([2, 0], [2, 0], None, -1., 3, [1., -1., 1.]),
+    # 3 classes. Only 0 and 1 are present, and are perfectly predicted. 2 should get absent_score.
+    pytest.param([0, 1], [0, 1], None, -1., 3, [1., 1., -1.]),
+    pytest.param([1, 0], [1, 0], None, -1., 3, [1., 1., -1.]),
+    # 3 classes, class 0 is 0.5 IoU, class 1 is 0 IoU (in pred but not target; should not get absent_score), class
+    # 2 is absent.
+    pytest.param([0, 1], [0, 0], None, -1., 3, [0.5, 0., -1.]),
+    # 3 classes, class 0 is 0.5 IoU, class 1 is 0 IoU (in target but not pred; should not get absent_score), class
+    # 2 is absent.
+    pytest.param([0, 0], [0, 1], None, -1., 3, [0.5, 0., -1.]),
+    # Sanity checks with absent_score of 1.0.
+    pytest.param([0, 2], [0, 2], None, 1.0, 3, [1., 1., 1.]),
+    pytest.param([0, 2], [0, 2], 0, 1.0, 3, [1., 1.]),
+])
 def test_iou_absent_score(pred, target, ignore_index, absent_score, num_classes, expected):
     iou_val = iou(
         pred=torch.tensor(pred),
@@ -413,29 +352,26 @@ def test_iou_absent_score(pred, target, ignore_index, absent_score, num_classes,
         ignore_index=ignore_index,
         absent_score=absent_score,
         num_classes=num_classes,
-        reduction="none",
+        reduction='none',
     )
     assert torch.allclose(iou_val, torch.tensor(expected).to(iou_val))
 
 
 # example data taken from
 # https://github.com/scikit-learn/scikit-learn/blob/master/sklearn/metrics/tests/test_ranking.py
-@pytest.mark.parametrize(
-    ["pred", "target", "ignore_index", "num_classes", "reduction", "expected"],
-    [
-        # Ignoring an index outside of [0, num_classes-1] should have no effect.
-        pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], None, 3, "none", [1, 1 / 2, 2 / 3]),
-        pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], -1, 3, "none", [1, 1 / 2, 2 / 3]),
-        pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 255, 3, "none", [1, 1 / 2, 2 / 3]),
-        # Ignoring a valid index drops only that index from the result.
-        pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 0, 3, "none", [1 / 2, 2 / 3]),
-        pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 1, 3, "none", [1, 2 / 3]),
-        pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 2, 3, "none", [1, 1 / 2]),
-        # When reducing to mean or sum, the ignored index does not contribute to the output.
-        pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 0, 3, "elementwise_mean", [7 / 12]),
-        pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 0, 3, "sum", [7 / 6]),
-    ],
-)
+@pytest.mark.parametrize(['pred', 'target', 'ignore_index', 'num_classes', 'reduction', 'expected'], [
+    # Ignoring an index outside of [0, num_classes-1] should have no effect.
+    pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], None, 3, 'none', [1, 1 / 2, 2 / 3]),
+    pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], -1, 3, 'none', [1, 1 / 2, 2 / 3]),
+    pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 255, 3, 'none', [1, 1 / 2, 2 / 3]),
+    # Ignoring a valid index drops only that index from the result.
+    pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 0, 3, 'none', [1 / 2, 2 / 3]),
+    pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 1, 3, 'none', [1, 2 / 3]),
+    pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 2, 3, 'none', [1, 1 / 2]),
+    # When reducing to mean or sum, the ignored index does not contribute to the output.
+    pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 0, 3, 'elementwise_mean', [7 / 12]),
+    pytest.param([0, 1, 1, 2, 2], [0, 1, 2, 2, 2], 0, 3, 'sum', [7 / 6]),
+])
 def test_iou_ignore_index(pred, target, ignore_index, num_classes, reduction, expected):
     iou_val = iou(
         pred=torch.tensor(pred),