+++ tests/trainer/logging_tests/test_eval_loop_logging_1_0.py
@@ -472,8 +472,6 @@ def test_log_works_in_val_callback(tmpdir):
                         "forked": False,
                         "func_name": func_name}
 
-        """
-
         def on_validation_start(self, trainer, pl_module):
             self.make_logging(pl_module, 'on_validation_start', 1, on_steps=self.choices,
                               on_epochs=self.choices, prob_bars=self.choices)
@@ -486,6 +484,7 @@ def test_log_works_in_val_callback(tmpdir):
             self.make_logging(pl_module, 'on_validation_epoch_start', 3, on_steps=self.choices,
                               on_epochs=self.choices, prob_bars=self.choices)
 
+        """
         def on_batch_start(self, trainer, pl_module):
             self.make_logging(pl_module, 'on_batch_start', 4, on_steps=self.choices,
                               on_epochs=self.choices, prob_bars=self.choices)
@@ -493,6 +492,7 @@ def test_log_works_in_val_callback(tmpdir):
         def on_validation_batch_start(self, trainer, pl_module, batch, batch_idx, dataloader_idx):
             self.make_logging(pl_module, 'on_validation_batch_start', 5, on_steps=self.choices,
                               on_epochs=self.choices, prob_bars=self.choices)
+        """
 
         def on_batch_end(self, trainer, pl_module):
             self.make_logging(pl_module, 'on_batch_end', 6, on_steps=self.choices,
@@ -509,8 +509,6 @@ def test_log_works_in_val_callback(tmpdir):
         def on_epoch_end(self, trainer, pl_module):
             self.make_logging(pl_module, 'on_epoch_end', 8, on_steps=[False],
                               on_epochs=self.choices, prob_bars=self.choices)
-
-        """
 
         def on_validation_epoch_end(self, trainer, pl_module):
             self.make_logging(pl_module, 'on_validation_epoch_end', 9, on_steps=[False],
@@ -541,16 +539,14 @@ def test_log_works_in_val_callback(tmpdir):
     trainer.fit(model)
     trainer.test()
 
-    """
     assert test_callback.funcs_called_count["on_epoch_start"] == 1
-    assert test_callback.funcs_called_count["on_batch_start"] == 1
+    # assert test_callback.funcs_called_count["on_batch_start"] == 1
     assert test_callback.funcs_called_count["on_batch_end"] == 1
     assert test_callback.funcs_called_count["on_validation_start"] == 1
     assert test_callback.funcs_called_count["on_validation_epoch_start"] == 1
-    assert test_callback.funcs_called_count["on_validation_batch_start"] == 4
+    # assert test_callback.funcs_called_count["on_validation_batch_start"] == 4
     assert test_callback.funcs_called_count["on_validation_batch_end"] == 4
     assert test_callback.funcs_called_count["on_epoch_end"] == 1
-    """
     assert test_callback.funcs_called_count["on_validation_epoch_end"] == 1
 
     # Make sure the func_name exists within callback_metrics. If not, we missed some
@@ -662,7 +658,6 @@ def test_log_works_in_test_callback(tmpdir):
                         "forked": False,
                         "func_name": func_name}
 
-        """
         def on_test_start(self, trainer, pl_module):
             self.make_logging(pl_module, 'on_test_start', 1, on_steps=self.choices,
                               on_epochs=self.choices, prob_bars=self.choices)
@@ -675,11 +670,8 @@ def test_log_works_in_test_callback(tmpdir):
             self.make_logging(pl_module, 'on_test_epoch_start', 3, on_steps=self.choices,
                               on_epochs=self.choices, prob_bars=self.choices)
 
-        def on_test_batch_start(self, trainer, pl_module, batch, batch_idx, dataloader_idx):
-            self.make_logging(pl_module, 'on_test_batch_start', 4, on_steps=self.choices,
-                              on_epochs=self.choices, prob_bars=self.choices)
-        def on_test_step_end(self, trainer, pl_module, outputs, batch, batch_idx, dataloader_idx):
-            self.make_logging(pl_module, 'on_test_step_end', 5, on_steps=self.choices,
+        def on_test_batch_end(self, trainer, pl_module, outputs, batch, batch_idx, dataloader_idx):
+            self.make_logging(pl_module, 'on_test_batch_end', 5, on_steps=self.choices,
                               on_epochs=self.choices, prob_bars=self.choices)
 
             # used to make sure aggregation works fine.
@@ -690,7 +682,6 @@ def test_log_works_in_test_callback(tmpdir):
         def on_epoch_end(self, trainer, pl_module):
             self.make_logging(pl_module, 'on_epoch_end', 6, on_steps=[False],
                               on_epochs=self.choices, prob_bars=self.choices)
-        """
 
         def on_test_epoch_end(self, trainer, pl_module):
             self.make_logging(pl_module, 'on_test_epoch_end', 7, on_steps=[False],
@@ -728,13 +719,11 @@ def test_log_works_in_test_callback(tmpdir):
     )
     trainer.fit(model)
     trainer.test()
-    """
+
     assert test_callback.funcs_called_count["on_test_start"] == 1
     assert test_callback.funcs_called_count["on_epoch_start"] == 2
     assert test_callback.funcs_called_count["on_test_epoch_start"] == 1
-    assert test_callback.funcs_called_count["on_test_batch_start"] == 4
-    assert test_callback.funcs_called_count["on_test_step_end"] == 4
-    """
+    assert test_callback.funcs_called_count["on_test_batch_end"] == 4
     assert test_callback.funcs_called_count["on_test_epoch_end"] == 1
 
     # Make sure the func_name exists within callback_metrics. If not, we missed some